Switch from ghc-proofs to inspection-testing (#15)
* Switch from ghc-proofs to inspection-testing

While ghc-proofs is a nice hack, it is actually unsuitable for
this kind of testing, as it optimizes the terms differently than GHC
(and more aggressivley). Therefore I have created the inspection-testing
package, which _really_ checks the output of GHC’s optimization
pipeline, and is equivalent to manually checking the output of
`-ddump-simpl`.

It turns out that the tests in the test suite do not pass with
GHC-8.0.2; I am not sure why that is. Insufficient inlining? CSE
failure? With GHC-8.2 we get the desired result.

I use CPP to disable the tests that do not work on GHC-8.0. If you can
get it to work again, then just remove the `#if` stanzas.

* Mention inspection-testing in the stack.yaml file