Make compilation faster by not going through an intermediate HList
In the previous version, the `HasField`, `HasType` and `HasPosition`
lenses were implemented using the same underlying class, `GHasKey`, which
first flattens the type out to a heterogeneous list, then uniformly
indexes into that list. This transformation (and especially optimising
it away) takes a significant time.

Instead, we can do this directly on the generic tree by providing a
predicate that determines which field to focus on. This results in
faster compile times, as the number of inlining passes required for
optimising away the overhead is greatly reduced.